Split macro into modules
